Get Rid of the discard mount option! see below.
So, SSDs are fast at writing, but only when their free space is neatly trimmed.
The only component in your software stack that knows which parts of your SSD should be trimmed,
is your file system. That is why there is a file system option in ext4 (my current file system of choice),
called .discard.. When this option is active, space that is freed up in the file system is reported to the SSD immediately,
and then the SSD does the trimming right away.
This will make the next write to that part of the SSD as fast as expected.
Obviously, trimming takes time . but how much time exactly?
I wanted to find out, and did the following:
I measured the time to unpack and then delete the kernel sources
(36706 files amounting to 493 MB, which is what I call a big bunch of small files).
I did it three times with and three times without the .discard. option,
and then took the average of those three tries:

Without .discard. option:

    Unpack: 1.21s
    Sync: 1.66s (= 172 MB/s)
    Delete: 0.47s
    Sync: 0.17s

With .discard. option:

    Unpack: 1.18s
    Sync: 1.62s (= 176 MB/s)
    Delete: 0.48s
    Sync: 40.41s

So, with .discard. on, deleting a big bunch of small files is 64 times slower on my SSD. For those ~40 seconds any I/O is really slow, so that.s pretty much the time when you get a fresh cup of coffee, or waste time watching the mass storage activity LED.

Conclusion
Don.t enable the .discard. option if you have a similar SSD. A much better way to keep your free space neatly trimmed for good write speeds is, to trigger a complete walk over the file system.s free space, and tell the SSD to trim that all at once. And of course you would do that at times when you don.t actually want to use the system (e.g. in a nightly cron job, or with a script that gets launched during system shutdown). This can be done with the .fstrim. command (that comes with util-linux), which takes around six minutes for my currently 60% filled 95 GB file system.