Major rebuild to all mounting process. +restored jornal to cache! I have found that system dalvik is in /cache and we must protect it! so i have restored data ordered journal back, and added check if no jornal found.
also we must not check filesystems every boot, not even cache!
we have ext4 journal, it's take care of all.

for second boot roms we have LOOP image, now this one we need to check before mount!
so i have left the INIT check to be used only for second boot ROMS.

also removed ro.kernel.android.checkjni 0 it's set by auto in ROM
no need to force it off. it's used for debug by VM to fix bad code apps.
so we need it.