Cortex Fixing, Logic and white space clean/fix.
WIFI now will start faster, if fource used.
Logger will reg the wakeup better.
Gestures and Vibrate and all fixing will be in BOOST range, not after it,
so no dealys! we dont like to WAIT! :)
if cortex proccess is OFF then load cpu tweaks and battery tweaks.
if ON will be loaded by function.