+ use more "local variable" [need testing]
"A variable declared as local is one that is visible only within the block of code in which it appears. It has local scope. In a function, a local variable has meaning only within that function block.
However, as Thomas Braunberger points out, a local variable declared in a function is also visible to functions called by the parent function." - http://tldp.org/LDP/abs/html/localvar.html