Added CPU OC bar from 1200 to 1600 by 100mhz step + info
Now to get OC speed, we need to set normal speed bar to 1200
then we can use the OC bar to change higher freq.
if user set OC to 1500 and normal to 1000 he get 1000!
but if he will set normal to 1200, he will get 1500 as OC bar setting!

also added to wake and boost functions in cortex.

in KSM, removed while function.
we call for KSM from wake function.