Set freq_cpu1on and freq_cpu1off for nightmare READ MORE: when using nightmare gov, it's very fast gov, so we need to control the sec cpu ON, or it's will be ON on each touch.
and we have max speed of 600Mhz at sleep, so it's best to set sec
cpu ON to that value, to help with wakeup.

when device sleep and awake to do some thing but screen is OFF
it's moving from 100~400Mhz so sec core should not turn on and waste
power.