New BUG FIX + New Function to check call state. + voltage fix
Added +25mV to 852/900/952 freq to prevent sod on call.
Reports show that some users get SOD on call or after it.
it's can be releated to voltage for cpu freqs that connected to radio
operations.

+ fixed root check.
now check if root app is also in /data/app

+ added call state check that will check if user on call, if yes then
no sleep function run, there is no need.

+ fixed stuck half sleep mode if power connected! awake was not running!
until no power source connected and screen off!

+ tuned RAM and removed sched options that can be unstable.

+ added force touch keys state from STweaks profile on boot
to set it OFF if user want that or fix the time out, as set by user.