Update root_check
In rootbox >3.9, we use the opensource su binary. The SuperUser app is integrated in ROM, so SuperUser from Chainfire is not needed anymore.
If the user has the root binary installed and it is executable and has SUID bit set, then the phone is rooted. It's just a problem of what app uses the binary, but the phone is considered to be rooted since su command will not fail. :)