Fixed Low Mem Kill tuning! here was critical mistake!
Every value that was set here is not the real ram size to be free for each group!

it's value * 4 = real MB limit! as it's set in lowmemorykiller.c

so when i say 2048 it's 2048 * 4 = 8MB RAM limit!

and so on! now all will work as should! and no more OOM dead phone.(SOD)