FIX damm profiles reset after boot. READ MORE!
For many users somehow the checksum that i check,
is changed in data partition after reboot!

And this lead to profile reset.

Now, for me, and many others all OK, but for some it's still bugged.

So i have removed the check, and now users need to READ the change log
and if i ask to reset the profile, and no one care!

Then, i dont care too.

end this shit.